API

IntervalConstraintProgramming.Vol — Type.

N-dimensional Volume with lower and upper bounds

source

IntervalConstraintProgramming.pave — Function.

pave(S::Separator, domain::IntervalBox, eps)`

Find the subset of domain defined by the constraints specified by the separator S. Returns (sub)pavings inner and boundary, i.e. lists of IntervalBox.

source

IntervalConstraintProgramming.pave — Method.

pave takes the given working list of boxes and splits them into inner and boundary lists with the given separator

source

IntervalConstraintProgramming.refine! — Function.

Refine a paving to tolerance ϵ

source

IntervalConstraintProgramming.CombinationSeparator — Type.

CombinationSeparator is a separator that is a combination (union, intersection, or complement) of other separators.

source

IntervalConstraintProgramming.ConstraintFunction — Type.

A ConstraintFunction contains the created forward and backward code

source

IntervalConstraintProgramming.ConstraintSeparator — Type.

ConstraintSeparator is a separator that represents a constraint defined directly using @constraint.

source

IntervalConstraintProgramming.Contractor — Type.

Contractor represents a Contractor from $\mathbb{R}^N$ to $\mathbb{R}^N$. Nout is the output dimension of the forward part.

source

Base.:∩ — Method.

∩(S1::Separator, S2::Separator)

Separator for the intersection of two sets given by the separators S1 and S2. Takes an iterator of intervals (IntervalBox, tuple, array, etc.), of length equal to the total number of variables in S1 and S2; it returns inner and outer tuples of the same length

source

IntervalConstraintProgramming.flatten! — Method.

flatten! recursively converts a Julia expression into a "flat" (one-dimensional) structure, stored in a FlatAST object. This is close to SSA (single-assignment form, https://en.wikipedia.org/wiki/Static_single_assignment_form).

Variables that are found are considered input_variables. Generated variables introduced at intermediate nodes are stored in intermediate. Returns the variable at the top of the current piece of the tree.

source

IntervalConstraintProgramming.isuniqued — Method.

Check if a symbol like :a has been uniqued to :_a_1_

source

IntervalConstraintProgramming.make_forward_function — Method.

Generate code for an anonymous function with given input arguments, output arguments, and code block.

source

IntervalConstraintProgramming.make_symbol — Method.

Return a new, unique symbol like _z3_

source

IntervalConstraintProgramming.mul_rev_new — Method.

∘_rev1(b, c, x) is the subset of x such that x ∘ b is defined and in c ∘_rev2(a, c, x) is the subset of x such that a ∘ x is defined and in c

If these agree (∘ is commutative) then call it ∘_rev(b, c, x)

source

IntervalConstraintProgramming.parse_comparison — Method.

parse_comparison parses comparisons like x >= 10 into the corresponding interval, expressed as x ∈ [10,∞]

Returns the expression and the constraint interval

TODO: Allow something like [3,4]' for the complement of [3,4]

source

IntervalConstraintProgramming.process_assignment! — Method.

An assigment is of the form a = f(...). The name a is currently retained. TODO: It should later be made unique.

source

IntervalConstraintProgramming.process_block! — Method.

A block represents a linear sequence of Julia statements. They are processed in order.

source

IntervalConstraintProgramming.process_call! — Function.

A call is something like +(x, y). A new variable is introduced for the result; its name can be specified using the new_var optional argument. If none is given, then a new, generated name is used.

source

IntervalConstraintProgramming.process_iterated_function! — Method.

Processes something of the form (f↑4)(x) (write as \uparrow<TAB>) by rewriting it to the equivalent set of iterated functions

source

IntervalConstraintProgramming.unify_variables — Method.

Unify the variables of two separators

source